refactor(generator): move password-generator into src/lib/generator/
Phase B of directory-structure-review. password-generator.ts is the weakest
src/lib root pin (stateless, single importer, no CI/CODEOWNERS coupling). Move
it next to its sibling generator-prefs.ts via the phase-config codemod.

- moves: src/lib/password-generator.{ts,test.ts} -> src/lib/generator/
- codemod auto-rewrote: the one importer (api/passwords/generate/route.ts),
  the file's own relative imports, and the vitest.config.ts coverage path
- removed password-generator.ts from the CONTRIBUTING.md src/lib pin list
  (10 -> 9 files) and updated the CLAUDE.md path reference, same commit
